+++ bin/check-haproxy.rb
@@ -98,6 +98,10 @@ class CheckHAProxy < Sensu::Plugin::Check::CLI
          short: '-m',
          boolean: true,
          description: 'Missing OK, flag enables'
+  option :missing_fail,
+         short: '-f',
+         boolean: false,
+         description: 'Missing CRIT, flag enables'
   option :service,
          short: '-s SVC',
          description: 'Service Name to Check'
@@ -115,7 +119,9 @@ class CheckHAProxy < Sensu::Plugin::Check::CLI
 
     if services.empty?
       message "No services matching /#{config[:service]}/"
-      if config[:missing_ok]
+      if config[:missing_fail]
+        critical
+      elsif config[:missing_ok]
         ok
       else
         warning